[SOL] Glue the insn setting $r5 to other in-reg passed args insns
The actual parameters to a function call are passed in registers for
the first 4 arguments, and on stack for the remaining arguments. The
register $r5 is set to the value of the frame pointer where the actual
parameters are stored. The instruction that copies the frame pointer
into the register $r5 must be glued to the chain of instructions that
sets the previous four call arguments. Otherwise, the scheduler is
free to move the instruction across boundaries of basic blocks and
it's potentially possible that the copy instruction is considered dead
code, even though there's an implicit liveness associated with the
call instruction.